moved modal into shared folder --- how do we make this component reusable? Copy and paste into our respective components folder and modify modal contents?
