+++ client/src/components/App.jsx
@@ -53,7 +53,7 @@ class App extends React.Component {
   // R&R HTTP requests
 
   render() {
-    const { products } = this.state;
+    const { products, isProductsLoaded } = this.state;
     return (
       <div>
         <h1>Hello World!</h1>
@@ -64,12 +64,14 @@ class App extends React.Component {
         <div id="RR">
           <RatingReview />
         </div>
-        <div id="QA">
-          <Form />
-          <Unit products={products} currentProduct="this.state currentProduct" />
-          {/*  current product id will determine what Q&A is displayed  */}
-          {/*  I will grab this product_id from whatever current display is :D */}
-        </div>
+        {isProductsLoaded ? (
+          <div id="QA">
+            <Form />
+            <Unit products={products} currentProduct="this.state currentProduct" />
+            {/*  current product id will determine what Q&A is displayed  */}
+            {/*  I will grab this product_id from whatever current display is :D */}
+          </div>
+        ) : ''}
       </div>
     );
   }
